src/requirements.txt
matplotlib>=3.7.1
numpy==1.26.4
Pillow==11.0.0
scikit_learn==1.3.0
torch==1.12.1
torchvision==0.13.1
torchaudio==0.12.1
tqdm==4.65.0
scipy==1.11.2
folktables
seaborn>=0.12.2